refactor: rename two_leg.py to spreads.py for clarity
