+++ optopsy/strategies/spreads.py
